fix(knowledge-network): optimize validation workspace
