+++ src/modules/knowledge-network/locales/en-US/network.ts
@@ -111,7 +111,12 @@ export const networkPart = {
     previewColIndexState: "Resource index state",
     previewUnbound: "Unbound",
     experienceTitle: "Capability verification",
+    experienceAgentTitle: "Q&A verification",
+    experienceMcpTitle: "MCP integration",
+    experienceRestTitle: "REST integration",
     experienceDescription: "Ask the knowledge network in natural language; answers are grounded in ontology retrieval with the retrieval trace shown.",
+    experienceMcpDescription: "Connect to knowledge network capabilities through MCP, inspect tools, and verify tool calls.",
+    experienceRestDescription: "Connect to knowledge network capabilities through REST, configure requests, and inspect live responses.",
     previewCreateObjectType: "Create object type",
     previewEdgeCount: "{{count}} relations",
     previewFitView: "Fit to view",
@@ -138,6 +143,9 @@ export const networkPart = {
     updatedBy: "Updated by: {{name}}",
     workspaceActionTypes: "Action types",
     workspaceAbilityVerification: "Capability verification",
+    workspaceExperienceAgent: "Q&A verification",
+    workspaceExperienceMcp: "MCP integration",
+    workspaceExperienceRest: "REST integration",
     workspaceConceptGroups: "Concept groups",
     workspaceDescription:
       "The workspace shell is already in place. The remaining legacy Vega pages will be migrated in later slices.",
