+++ src/modules/knowledge-network/locales/zh-CN/network.ts
@@ -112,7 +112,12 @@ export const networkPart = {
     previewColIndexState: "资源索引状态",
     previewUnbound: "未绑定",
     experienceTitle: "能力验证",
+    experienceAgentTitle: "智能问答",
+    experienceMcpTitle: "MCP 对接",
+    experienceRestTitle: "REST 对接",
     experienceDescription: "用自然语言向知识网络提问，基于本体检索作答并展示检索过程。",
+    experienceMcpDescription: "通过 MCP 协议对接知识网络能力，查看工具列表并验证工具调用。",
+    experienceRestDescription: "通过 REST 接口对接知识网络能力，配置请求参数并查看实时响应。",
     recentlyModifiedObjectTypes: "最近修改的对象类",
     resourcePageExit: "退出",
     closeSearchBar: "关闭搜索栏",
@@ -132,6 +137,9 @@ export const networkPart = {
     updatedBy: "更新人：{{name}}",
     workspaceActionTypes: "行动类",
     workspaceAbilityVerification: "能力验证",
+    workspaceExperienceAgent: "智能问答",
+    workspaceExperienceMcp: "MCP 对接",
+    workspaceExperienceRest: "REST 对接",
     workspaceConceptGroups: "概念分组",
     workspaceDescription:
       "工作区壳层已经迁入，后续会继续补齐旧 Vega 的完整业务页。",
