+++ src/modules/knowledge-network/routes/lazy-pages.tsx
@@ -23,11 +23,6 @@ export const KnowledgeNetworkListPage = lazyNamedExport(
   "KnowledgeNetworkListPage",
 );
 
-export const ExperiencePage = lazyNamedExport(
-  () => import("@/modules/knowledge-network/pages/ExperiencePage"),
-  "ExperiencePage",
-);
-
 export const KnowledgeNetworkWorkspaceStandalonePage = lazy(async () => {
   const module = await import(
     "@/modules/knowledge-network/pages/KnowledgeNetworkWorkspaceStandalonePage"
