+++ src/modules/knowledge-network/scenes/ExperienceScene.tsx
@@ -12,16 +12,12 @@
 
 import {
   ArrowLeftOutlined,
-  ApiOutlined,
   CopyOutlined,
-  DatabaseOutlined,
   KeyOutlined,
   QuestionCircleOutlined,
-  ReadOutlined,
-  ThunderboltFilled,
 } from "@ant-design/icons";
-import { App, Empty, Input, Modal, Segmented, Select, Spin, Tabs, Tooltip } from "antd";
-import { useCallback, useEffect, useMemo, useRef, useState, type ReactNode } from "react";
+import { App, Select, Tooltip } from "antd";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 
 import { gatewayOrigin } from "@/framework/auth/oauth";
@@ -37,7 +33,6 @@ import {
   exampleBodyText,
   fetchKnDetail,
   fetchObjectInstances,
-  opSupportsTestData,
   pickQueryableObjectType,
   listMcpTools,
   mcpPathOf,
@@ -50,32 +45,19 @@ import {
   type KnDetail,
   type KnObjectType,
   type KnRelationType,
-  type McpAuth,
   type McpToolDef,
 } from "@/modules/knowledge-network/services/context-loader.service";
 import { AgentChat } from "@/modules/knowledge-network/components/agent-chat/AgentChat";
 import type { AgentTokenProvider } from "@/modules/knowledge-network/services/agent-chat.service";
+import { ContextLoaderIntegrationPanel } from "./ContextLoaderIntegrationPanel";
+import { DataBrowserPanel } from "./DataBrowserPanel";
+import { McpSetupModal, ToolDiscoveryModal } from "./McpIntegrationModals";
 
 import styles from "./ExperienceScene.module.css";
 
 /** 线上有、本地无 op 定义的工具（如 get_object_types / get_relation_types）归到 Knowledge Network 组，不单开分类。 */
 const ONLINE_GROUP = "Knowledge Network";
 
-function formatPreviewValue(value: unknown) {
-  if (value === null || value === undefined) {
-    return "—";
-  }
-  if (typeof value === "string" || typeof value === "number" || typeof value === "boolean") {
-    return String(value);
-  }
-  try {
-    return JSON.stringify(value) ?? "";
-  } catch {
-    return "";
-  }
-}
-
-/** 从单个 JSON Schema 属性生成可编辑示例值。 */
 function sampleForSchemaProp(def: unknown): unknown {
   if (!def || typeof def !== "object") return "";
   const d = def as Record<string, unknown>;
@@ -123,13 +105,6 @@ function synthesizeOp(tool: McpToolDef): ContextLoaderOp {
     body,
     mcpArgs: body,
   };
-}
-
-/** 字节数转人类可读：B / KB / MB（1024 进制，保留一位小数）。 */
-function formatBytes(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
 }
 
 /** 从 MCP tools/call 信封里抽出 result.content[].text（TOON 或 JSON 文本载荷）。 */
@@ -192,66 +167,6 @@ function findArrayProp(node: unknown, key: string): unknown[] | null {
 }
 
 /* ============================ JSON 语法高亮（无依赖，正则分词） ============================ */
-const JSON_TOKEN_RE = /("(?:\\.|[^"\\])*")(\s*:)?|\b(true|false|null)\b|(-?\d+(?:\.\d+)?(?:[eE][+-]?\d+)?)/g;
-
-function JsonHighlight({ text }: { text: string }) {
-  // 超大响应不逐 token 渲染，避免卡顿。
-  if (text.length > 200_000) return <>{text}</>;
-  const nodes: ReactNode[] = [];
-  let last = 0;
-  let key = 0;
-  JSON_TOKEN_RE.lastIndex = 0;
-  let m: RegExpExecArray | null;
-  while ((m = JSON_TOKEN_RE.exec(text)) !== null) {
-    if (m.index > last) nodes.push(text.slice(last, m.index));
-    if (m[1] !== undefined) {
-      if (m[2] !== undefined) {
-        // 字符串后紧跟冒号 → 属性名（key）
-        nodes.push(<span key={key++} className={styles.jKey}>{m[1]}</span>);
-        nodes.push(<span key={key++} className={styles.jPunct}>{m[2]}</span>);
-      } else {
-        nodes.push(<span key={key++} className={styles.jStr}>{m[1]}</span>);
-      }
-    } else if (m[3] !== undefined) {
-      nodes.push(<span key={key++} className={styles.jKw}>{m[3]}</span>);
-    } else if (m[4] !== undefined) {
-      nodes.push(<span key={key++} className={styles.jNum}>{m[4]}</span>);
-    }
-    last = JSON_TOKEN_RE.lastIndex;
-  }
-  if (last < text.length) nodes.push(text.slice(last));
-  return <>{nodes}</>;
-}
-
-/* ============================ 可编辑 JSON 编辑器（透明 textarea + 背后高亮 pre，滚动同步） ============================ */
-function JsonEditor({ value, onChange }: { value: string; onChange: (next: string) => void }) {
-  const taRef = useRef<HTMLTextAreaElement>(null);
-  const preRef = useRef<HTMLPreElement>(null);
-  const syncScroll = () => {
-    if (preRef.current && taRef.current) {
-      preRef.current.scrollTop = taRef.current.scrollTop;
-      preRef.current.scrollLeft = taRef.current.scrollLeft;
-    }
-  };
-  return (
-    <div className={styles.editWrap}>
-      <pre ref={preRef} className={styles.editHl} aria-hidden="true">
-        <JsonHighlight text={value} />
-        {"\n"}
-      </pre>
-      <textarea
-        ref={taRef}
-        className={styles.ta}
-        value={value}
-        spellCheck={false}
-        onChange={(event) => onChange(event.target.value)}
-        onScroll={syncScroll}
-      />
-    </div>
-  );
-}
-
-/* ============================ API Key 掩码输入（失焦掩码头+尾，聚焦显全编辑） ============================ */
 function maskKey(value: string): string {
   const v = value.trim();
   return v.length <= 12 ? v : `${v.slice(0, 8)}****${v.slice(-4)}`;
@@ -298,621 +213,17 @@ function MaskedKeyInput({
 }
 
 /* ============================ MCP 接入指南（Claude Code / Cursor / 通用） ============================ */
-function CodeBlock({
-  title,
-  code,
-  json,
-  onCopy,
-}: {
-  title: string;
-  code: string;
-  json?: boolean;
-  onCopy: () => void;
-}) {
-  return (
-    <div className={styles.codeBlk}>
-      <div className={styles.codeBlkHead}>
-        <span>{title}</span>
-        <button type="button" className={styles.mini} onClick={onCopy}>
-          <CopyOutlined /> 复制
-        </button>
-      </div>
-      <pre className={styles.codeBlkPre}>{json ? <JsonHighlight text={code} /> : code}</pre>
-    </div>
-  );
-}
+type ExperienceSceneProps = {
+  embedded?: boolean;
+  initialMode?: ContextLoaderMode;
+  lockMode?: boolean;
+};
 
-function McpSetupModal({
-  open,
-  onClose,
-  mcpUrl,
-  onIssueKey,
-  copy,
-}: {
-  open: boolean;
-  onClose: () => void;
-  mcpUrl: string;
-  onIssueKey: () => void;
-  copy: (text: string, label?: string) => void;
-}) {
-  const tk = "bak_<在「API Key」页签发的长期 Key>";
-  const jsonConfig = JSON.stringify(
-    {
-      mcpServers: {
-        "bkn-agent-retrieval": {
-          type: "http",
-          url: mcpUrl,
-          headers: { Authorization: `Bearer ${tk}` },
-        },
-      },
-    },
-    null,
-    2,
-  );
-  const claudeCli = [
-    `claude mcp add --transport http bkn-agent-retrieval ${mcpUrl} \\`,
-    `  --header "Authorization: Bearer ${tk}"`,
-  ].join("\n");
-
-  return (
-    <Modal open={open} onCancel={onClose} footer={null} width={680} title="接入 MCP（Claude Code / Cursor）">
-      <div className={styles.guideRoot}>
-      <p className={styles.guideNote}>
-        接入指南用于<b>外部 MCP 客户端 / SDK</b>（Cursor、Claude Code 等）。鉴权填 <b>AppKey</b>（<code>bak_</code> 开头的长期 Key），
-        在左下角「API Key」页签发。
-        <button type="button" className={styles.guideLink} onClick={onIssueKey}>
-          去签发 AppKey →
-        </button>
-      </p>
-      <p className={styles.guideNote}>
-        <b>和本页登录态的差异：</b>本页调试用的是你的<b>会话 token</b>（<code>ory_at_</code>，几十分钟就过期，只够即时调试）；
-        外部客户端要长期可用，必须用 <b>AppKey</b>（<code>bak_</code>，长期有效、可撤销、可轮换）。两者都放同一个
-        <code>Authorization: Bearer</code> 头，网关按前缀自动识别。
-      </p>
-      <Tabs
-        defaultActiveKey="claude"
-        items={[
-          {
-            key: "claude",
-            label: "Claude Code",
-            children: (
-              <>
-                <CodeBlock title="① CLI 一行接入" code={claudeCli} onCopy={() => copy(claudeCli, "命令已复制")} />
-                <CodeBlock
-                  title="② 或写入项目 .mcp.json"
-                  code={jsonConfig}
-                  json
-                  onCopy={() => copy(jsonConfig, "配置已复制")}
-                />
-              </>
-            ),
-          },
-          {
-            key: "cursor",
-            label: "Cursor",
-            children: (
-              <>
-                <p className={styles.guideNote}>
-                  写入 <code>~/.cursor/mcp.json</code>（全局）或项目内 <code>.cursor/mcp.json</code>，重启 Cursor 后生效。
-                </p>
-                <CodeBlock title="~/.cursor/mcp.json" code={jsonConfig} json onCopy={() => copy(jsonConfig, "配置已复制")} />
-              </>
-            ),
-          },
-          {
-            key: "generic",
-            label: "通用 (mcp.json)",
-            children: (
-              <CodeBlock title="mcpServers 配置" code={jsonConfig} json onCopy={() => copy(jsonConfig, "配置已复制")} />
-            ),
-          },
-        ]}
-      />
-      </div>
-    </Modal>
-  );
-}
-
-/* ============================ 工具发现（tools/list：动态发现 + 与本地硬编码漂移对照） ============================ */
-function SchemaPre({ title, value, copy }: { title: string; value: unknown; copy: (text: string, label?: string) => void }) {
-  const text = JSON.stringify(value ?? {}, null, 2);
-  return (
-    <div className={styles.toolSchema}>
-      <div className={styles.codeBlkHead}>
-        <span>{title}</span>
-        <button type="button" className={styles.mini} onClick={() => copy(text, `${title} 已复制`)}>
-          <CopyOutlined /> 复制
-        </button>
-      </div>
-      <pre className={styles.codeBlkPre}>
-        <JsonHighlight text={text} />
-      </pre>
-    </div>
-  );
-}
-
-function ToolDiscoveryModal({
-  open,
-  onClose,
-  tools,
-  loading,
-  error,
-  onReload,
-  copy,
-}: {
-  open: boolean;
-  onClose: () => void;
-  tools: McpToolDef[] | null;
-  loading: boolean;
-  error: string | null;
-  onReload: () => void;
-  copy: (text: string, label?: string) => void;
-}) {
-  return (
-    <Modal open={open} onCancel={onClose} footer={null} width={720} title="工具发现 · tools/list">
-      <div className={styles.guideRoot}>
-        <p className={styles.guideNote}>
-          直接向 MCP <code>tools/list</code> 拉取所有线上工具及其 <code>inputSchema</code> / <code>outputSchema</code>。
-          左侧 MCP 接口列表已<b>实时由 tools/list 驱动</b>（后端新增工具自动出现）；这里查看各工具完整 schema。
-          <button type="button" className={styles.guideLink} onClick={onReload}>
-            重新拉取 →
-          </button>
-        </p>
-        {loading ? (
-          <div className={styles.discoverEmpty}>
-            <Spin />
-          </div>
-        ) : error ? (
-          <div className={styles.resError}>
-            <ApiOutlined />
-            <div>
-              <strong>拉取失败</strong>
-              <p>{error}</p>
-            </div>
-          </div>
-        ) : tools ? (
-          <>
-            <div className={styles.driftRow}>
-              <span className={styles.driftStat}>线上 {tools.length} 个</span>
-            </div>
-            <div className={styles.toolList}>
-              {tools.map((tool) => (
-                <details key={tool.name} className={styles.toolItem}>
-                  <summary className={styles.toolSummary}>
-                    <span className={styles.toolName}>{tool.name}</span>
-                    {tool.description ? <span className={styles.toolDesc}>{tool.description}</span> : null}
-                  </summary>
-                  <SchemaPre title="inputSchema" value={tool.inputSchema} copy={copy} />
-                  {tool.outputSchema !== undefined ? (
-                    <SchemaPre title="outputSchema" value={tool.outputSchema} copy={copy} />
-                  ) : null}
-                </details>
-              ))}
-            </div>
-          </>
-        ) : null}
-      </div>
-    </Modal>
-  );
-}
-
-/* ============================ 数据浏览器（右侧抽屉：schema + 资源 id，点击填入请求体） ============================ */
-function ObjectTypeCard({
-  ot,
-  onFillField,
-  onFillResource,
-  onFillTest,
-  copy,
-  env,
-  auth,
-}: {
-  ot: KnObjectType;
-  onFillField: (key: string, value: string) => void;
-  onFillResource: (resourceId: string) => void;
-  /** 用该对象类型的真实样本行填充当前接口；仅在当前接口按对象类型取数时传入。 */
-  onFillTest?: (ot: KnObjectType) => Promise<void>;
-  copy: (text: string, label?: string) => void;
-  env: ContextLoaderEnv;
-  /** 401 自动刷新 token 用（OAuth 续期）。 */
-  auth?: McpAuth;
-}) {
-  const [open, setOpen] = useState(false);
-  const [filling, setFilling] = useState(false);
-  const res = ot.data_source ?? null;
-  const props = ot.data_properties ?? [];
-
-  // 样本行预览（按需拉取 query_object_instance）
-  const [previewOpen, setPreviewOpen] = useState(false);
-  const [previewRows, setPreviewRows] = useState<Record<string, unknown>[] | null>(null);
-  const [previewLoading, setPreviewLoading] = useState(false);
-  const [previewError, setPreviewError] = useState<string | null>(null);
-
-  const togglePreview = () => {
-    const next = !previewOpen;
-    setPreviewOpen(next);
-    if (next && previewRows === null && !previewLoading) {
-      setPreviewLoading(true);
-      setPreviewError(null);
-      fetchObjectInstances(env, ot.id, 5, auth)
-        .then((rows) => setPreviewRows(rows))
-        .catch((error) => setPreviewError(error instanceof Error ? error.message : "查询失败"))
-        .finally(() => setPreviewLoading(false));
-    }
-  };
-
-  const previewColumns =
-    props.length > 0
-      ? props.map((p) => p.name)
-      : previewRows && previewRows[0]
-        ? Object.keys(previewRows[0]).filter((k) => !k.startsWith("_"))
-        : [];
-
-  return (
-    <div className={styles.dbCard}>
-      <div className={styles.dbCardHead}>
-        <span className={styles.dbOtName} title={ot.name || ot.id}>
-          {ot.name || ot.id}
-        </span>
-        {onFillTest && res?.id ? (
-          <Tooltip title="用该对象类型的真实样本行填充当前接口请求体">
-            <button
-              type="button"
-              className={styles.dbTestBtn}
-              disabled={filling}
-              onClick={() => {
-                if (filling) return;
-                setFilling(true);
-                void onFillTest(ot).finally(() => setFilling(false));
-              }}
-            >
-              {filling ? <Spin size="small" /> : <ThunderboltFilled />} 填入测试请求
-            </button>
-          </Tooltip>
-        ) : null}
-        <button
-          type="button"
-          className={`${styles.dbFields} ${open ? styles.dbFieldsOpen : ""}`}
-          onClick={() => setOpen((value) => !value)}
-          disabled={props.length === 0}
-        >
-          {props.length} 字段 <span className={styles.dbChev}>▾</span>
-        </button>
-      </div>
-
-      <div className={styles.dbRow}>
-        <span className={styles.dbRowLabel}>对象类型</span>
-        <Tooltip title="点击填入当前接口的 ot_id">
-          <button type="button" className={styles.dbChip} onClick={() => onFillField("ot_id", ot.id)}>
-            {ot.id}
-          </button>
-        </Tooltip>
-        <Tooltip title="复制 ot_id">
-          <button type="button" className={styles.dbCopy} onClick={() => copy(ot.id, "已复制 ot_id")}>
-            <CopyOutlined />
-          </button>
-        </Tooltip>
-      </div>
-
-      <div className={styles.dbRow}>
-        <span className={styles.dbRowLabel}>数据资源</span>
-        {res?.id ? (
-          <>
-            <Tooltip title="点击填入 run_sql 的 {{资源}} 占位（其它接口则复制）">
-              <button type="button" className={styles.dbRes} onClick={() => onFillResource(res.id)}>
-                <DatabaseOutlined /> {res.name || "资源"} · {res.id}
-              </button>
-            </Tooltip>
-            <Tooltip title="复制资源 id">
-              <button type="button" className={styles.dbCopy} onClick={() => copy(res.id, "已复制资源 id")}>
-                <CopyOutlined />
-              </button>
-            </Tooltip>
-          </>
-        ) : (
-          <span className={styles.dbNoRes}>无绑定</span>
-        )}
-      </div>
-
-      {open && props.length > 0 ? (
-        <div className={styles.dbPropList}>
-          <div className={styles.dbPropHead}>字段 · 点击复制名称</div>
-          {props.map((prop) => (
-            <Tooltip key={prop.name} title={`复制字段名 ${prop.name}`}>
-              <button
-                type="button"
-                className={styles.dbProp}
-                onClick={() => copy(prop.name, `已复制 ${prop.name}`)}
-              >
-                <span className={styles.dbPropName}>{prop.name}</span>
-                {prop.display_name && prop.display_name !== prop.name ? (
-                  <span className={styles.dbPropDisp}>{prop.display_name}</span>
-                ) : null}
-                <span className={styles.dbPropType}>{prop.type || "—"}</span>
-                <CopyOutlined className={styles.dbPropCopy} />
-              </button>
-            </Tooltip>
-          ))}
-        </div>
-      ) : null}
-
-      <div className={styles.dbRow}>
-        <span className={styles.dbRowLabel}>样本数据</span>
-        <button
-          type="button"
-          className={`${styles.dbFields} ${previewOpen ? styles.dbFieldsOpen : ""}`}
-          onClick={togglePreview}
-        >
-          {previewOpen ? "收起预览" : "预览数据"} <span className={styles.dbChev}>▾</span>
-        </button>
-      </div>
-
-      {previewOpen ? (
-        <div className={styles.dbPreview}>
-          {previewLoading ? (
-            <div className={styles.dbPreviewMsg}>
-              <Spin size="small" /> 加载中…
-            </div>
-          ) : previewError ? (
-            <div className={styles.dbPreviewErr}>{previewError}</div>
-          ) : previewRows && previewRows.length > 0 && previewColumns.length > 0 ? (
-            <div className={styles.dbPreviewTableWrap}>
-              <table className={styles.dbPreviewTable}>
-                <thead>
-                  <tr>
-                    {previewColumns.map((col) => (
-                      <th key={col}>{col}</th>
-                    ))}
-                  </tr>
-                </thead>
-                <tbody>
-                  {previewRows.map((row, rowIndex) => (
-                    <tr key={rowIndex}>
-                      {previewColumns.map((col) => {
-                        const value = row[col];
-                        const text = formatPreviewValue(value);
-                        return (
-                          <td key={col} title={text}>
-                            {text}
-                          </td>
-                        );
-                      })}
-                    </tr>
-                  ))}
-                </tbody>
-              </table>
-            </div>
-          ) : (
-            <div className={styles.dbPreviewMsg}>无数据</div>
-          )}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function DataBrowserPanel({
-  active,
-  env,
-  knName,
-  onFillField,
-  onFillResource,
-  onFillConceptGroup,
-  onFillTest,
-  onFillRelation,
-  copy,
-  auth,
-}: {
-  active: boolean;
-  env: ContextLoaderEnv;
-  knName: string;
-  onFillField: (key: string, value: string) => void;
-  onFillResource: (resourceId: string) => void;
-  onFillConceptGroup: (groupId: string) => void;
-  /** 当前接口按对象类型取数时传入，使每张卡片可一键填充测试请求。 */
-  onFillTest?: (ot: KnObjectType) => Promise<void>;
-  /** 当前接口为 query_instance_subgraph 时传入，使关系卡可一键填入子图路径。 */
-  onFillRelation?: (rel: KnRelationType) => void;
-  copy: (text: string, label?: string) => void;
-  /** 401 自动刷新 token 用（OAuth 续期）。 */
-  auth?: McpAuth;
-}) {
-  const [detail, setDetail] = useState<KnDetail | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [q, setQ] = useState("");
-  const [view, setView] = useState<"object" | "relation">("object");
-  const loadedRef = useRef(false);
-
-  // 懒加载：首次切到「数据浏览器」标签时拉一次 schema，之后常驻不再重拉，保留预览/筛选上下文。
-  useEffect(() => {
-    if (!active || loadedRef.current) return;
-    loadedRef.current = true;
-    let cancelled = false;
-    setLoading(true);
-    setError(null);
-    fetchKnDetail(env, auth)
-      .then((data) => {
-        if (!cancelled) setDetail(data);
-      })
-      .catch((err) => {
-        if (!cancelled) {
-          setError(err instanceof Error ? err.message : "加载失败");
-          loadedRef.current = false; // 失败可重试
-        }
-      })
-      .finally(() => {
-        if (!cancelled) setLoading(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [active, env, auth]);
-
-  const sections = useMemo(() => {
-    if (!detail) return [];
-    const needle = q.trim().toLowerCase();
-    const match = (ot: KnObjectType) =>
-      !needle ||
-      `${ot.id} ${ot.name ?? ""} ${ot.data_source?.id ?? ""} ${ot.data_source?.name ?? ""}`
-        .toLowerCase()
-        .includes(needle);
-    const byId = new Map(detail.object_types.map((o) => [o.id, o]));
-    const grouped = detail.concept_groups.map((group) => ({
-      id: group.id,
-      title: group.name || group.id,
-      ots: (group.object_type_ids ?? []).map((oid) => byId.get(oid)).filter((o): o is KnObjectType => Boolean(o)),
-    }));
-    const inGroup = new Set(detail.concept_groups.flatMap((g) => g.object_type_ids ?? []));
-    const ungrouped = detail.object_types.filter((o) => !inGroup.has(o.id));
-    if (ungrouped.length) grouped.push({ id: "", title: "未分组", ots: ungrouped });
-    return grouped
-      .map((section) => ({ ...section, ots: section.ots.filter(match) }))
-      .filter((section) => section.ots.length > 0);
-  }, [detail, q]);
-
-  const relations = useMemo(() => {
-    if (!detail) return [];
-    const needle = q.trim().toLowerCase();
-    return detail.relation_types.filter(
-      (rel) =>
-        !needle ||
-        `${rel.id} ${rel.name ?? ""} ${rel.sourceId} ${rel.targetId}`.toLowerCase().includes(needle),
-    );
-  }, [detail, q]);
-
-  return (
-    <div className={styles.dbWrap}>
-      <div className={styles.dbTitle}>
-        <DatabaseOutlined /> 数据浏览器 · {knName || "知识网络"}
-      </div>
-        <p className={styles.dbHint}>
-          点「<b>+ 资源组</b>」→ 加入 <code>concept_groups</code>；点「对象类型」→ 填入当前接口的 <code>ot_id</code>；
-          点「数据资源」→ 填入 run_sql 的 <code>{"{{资源}}"}</code> 占位；点「预览数据」看样本行。
-        </p>
-        <div className={styles.dbSearch}>
-          <Input
-            value={q}
-            onChange={(e) => setQ(e.target.value)}
-            placeholder={view === "object" ? "筛选对象类型 / 资源…" : "筛选关系类…"}
-            allowClear
-          />
-        </div>
-        {detail && !loading && !error ? (
-          <div className={styles.dbViewSwitch}>
-            <Segmented
-              size="small"
-              block
-              value={view}
-              onChange={(value) => setView(value as "object" | "relation")}
-              options={[
-                { label: `对象类型 ${detail.object_types.length}`, value: "object" },
-                { label: `关系类 ${detail.relation_types.length}`, value: "relation" },
-              ]}
-            />
-          </div>
-        ) : null}
-        <div className={styles.dbList}>
-          {loading ? (
-            <div className={styles.dbCenter}>
-              <Spin />
-            </div>
-          ) : error ? (
-            <div className={styles.dbError}>
-              <ApiOutlined />
-              <div>
-                <strong>加载失败</strong>
-                <p>{error}</p>
-              </div>
-            </div>
-          ) : view === "object" ? (
-            sections.length === 0 ? (
-              <div className={styles.dbCenter}>
-                <Empty description="无匹配对象类型" />
-              </div>
-            ) : (
-              sections.map((section) => (
-                <div key={section.title} className={styles.dbSection}>
-                  {section.id ? (
-                    <div className={styles.dbGroupRow}>
-                      <span className={styles.dbGroup}>{section.title}</span>
-                      <Tooltip title={`加入 concept_groups：${section.id}`}>
-                        <button
-                          type="button"
-                          className={styles.dbGroupAdd}
-                          onClick={() => onFillConceptGroup(section.id)}
-                        >
-                          + 资源组
-                        </button>
-                      </Tooltip>
-                    </div>
-                  ) : (
-                    <div className={styles.dbGroup}>{section.title}</div>
-                  )}
-                  {section.ots.map((ot) => (
-                    <ObjectTypeCard
-                      key={ot.id}
-                      ot={ot}
-                      onFillField={onFillField}
-                      onFillResource={onFillResource}
-                      onFillTest={onFillTest}
-                      copy={copy}
-                      env={env}
-                      auth={auth}
-                    />
-                  ))}
-                </div>
-              ))
-            )
-          ) : relations.length === 0 ? (
-            <div className={styles.dbCenter}>
-              <Empty description="无匹配关系类" />
-            </div>
-          ) : (
-            <div className={styles.dbSection}>
-              {relations.map((rel) => (
-                <div key={rel.id} className={styles.dbCard}>
-                  <div className={styles.dbCardHead}>
-                    <span className={styles.dbOtName} title={rel.name || rel.id}>
-                      {rel.name || rel.id}
-                    </span>
-                    {onFillRelation ? (
-                      <Tooltip title="填入 query_instance_subgraph 的 relation_type_paths">
-                        <button type="button" className={styles.dbTestBtn} onClick={() => onFillRelation(rel)}>
-                          <ThunderboltFilled /> 填入子图
-                        </button>
-                      </Tooltip>
-                    ) : null}
-                  </div>
-                  <div className={styles.dbRow}>
-                    <span className={styles.dbRowLabel}>路径</span>
-                    <Tooltip title="点击填入 ot_id">
-                      <button type="button" className={styles.dbChip} onClick={() => onFillField("ot_id", rel.sourceId)}>
-                        {rel.sourceId}
-                      </button>
-                    </Tooltip>
-                    <span className={styles.dbRelArrow}>→</span>
-                    <Tooltip title="点击填入 ot_id">
-                      <button type="button" className={styles.dbChip} onClick={() => onFillField("ot_id", rel.targetId)}>
-                        {rel.targetId}
-                      </button>
-                    </Tooltip>
-                    <Tooltip title="复制关系 id">
-                      <button type="button" className={styles.dbCopy} onClick={() => copy(rel.id, "已复制关系 id")}>
-                        <CopyOutlined />
-                      </button>
-                    </Tooltip>
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      </div>
-  );
-}
-
-/* ============================ 主场景 ============================ */
-export function ExperienceScene() {
+export function ExperienceScene({
+  embedded = false,
+  initialMode = "agent",
+  lockMode = false,
+}: ExperienceSceneProps) {
   const navigate = useNavigate();
   const runtimeConfig = useRuntimeConfig();
   const { message } = App.useApp();
@@ -930,7 +241,13 @@ export function ExperienceScene() {
   );
 
   const [network, setNetwork] = useState<{ name: string; slug: string } | null>(null);
-  const [mode, setMode] = useState<ContextLoaderMode>("agent");
+  const [mode, setMode] = useState<ContextLoaderMode>(initialMode);
+  const showModeTabs = !lockMode;
+  const showEnvSettings = mode !== "agent";
+
+  useEffect(() => {
+    setMode(initialMode);
+  }, [initialMode]);
 
   // 请求基址：走当前源（dev 经 vite 代理转后端，避免浏览器跨域）。
   const [base] = useState(() => (typeof window !== "undefined" ? window.location.origin : "http://agent-retrieval:30779"));
@@ -969,7 +286,7 @@ export function ExperienceScene() {
   const [response, setResponse] = useState<ContextLoaderResponse | null>(null);
   const [reqError, setReqError] = useState<string | null>(null);
   const [sending, setSending] = useState(false);
-  const [curlOpen, setCurlOpen] = useState(true);
+  const [curlOpen, setCurlOpen] = useState(false);
   const [guideOpen, setGuideOpen] = useState(false);
   const [discoverOpen, setDiscoverOpen] = useState(false);
   const [rightTab, setRightTab] = useState<"res" | "data">("res");
@@ -1266,16 +583,16 @@ export function ExperienceScene() {
     [bodyText, copy, message],
   );
 
-  const verb = mode === "mcp" ? "MCP" : "POST";
-
   return (
-    <section className={styles.page}>
+    <section className={`${styles.page} ${embedded ? styles.pageEmbedded : ""}`}>
+      {showModeTabs || showEnvSettings ? (
       <div className={styles.topbar}>
-        {network ? (
+        {!embedded && network ? (
           <button type="button" className={styles.back} onClick={() => void navigate(`/knowledge-network/workspace/${id}/overview`)}>
             <ArrowLeftOutlined /> 返回 {network.name}
           </button>
         ) : null}
+        {showModeTabs ? (
         <div className={styles.tabs}>
           {(["agent", "mcp", "rest"] as ContextLoaderMode[]).map((value) => (
             <button
@@ -1288,6 +605,8 @@ export function ExperienceScene() {
             </button>
           ))}
         </div>
+        ) : null}
+        {showEnvSettings ? (
         <div className={styles.envset}>
           <div className={styles.ef}>
             <label>知识网络 kn_id</label>
@@ -1336,7 +655,9 @@ export function ExperienceScene() {
             </div>
           ) : null}
         </div>
+        ) : null}
       </div>
+      ) : null}
 
       {mode === "agent" ? (
         <AgentChat
@@ -1346,292 +667,66 @@ export function ExperienceScene() {
           modelTokenProvider={modelTokenProvider}
         />
       ) : (
-        <div className={styles.main}>
-          {/* 接口列表 */}
-          <aside className={styles.list}>
-            {mode === "mcp" ? (
-              <>
-                <button type="button" className={styles.guideBtn} onClick={() => setGuideOpen(true)}>
-                  <ReadOutlined /> 接入 Claude Code / Cursor
-                </button>
-                <button type="button" className={styles.discoverBtn} onClick={() => setDiscoverOpen(true)}>
-                  <ApiOutlined /> 工具发现 · tools/list
-                </button>
-              </>
-            ) : null}
-            <div className={styles.listSearch}>
-              <Input value={filter} onChange={(e) => setFilter(e.target.value)} placeholder="筛选接口…" />
-            </div>
-            <div className={styles.eplist}>
-              {[...new Set(activeOps.map((item) => item.group))].map((group) => {
-                const items = activeOps.filter(
-                  (item) =>
-                    item.group === group &&
-                    (!filter || (item.id + item.path + item.summary).toLowerCase().includes(filter.toLowerCase())),
-                );
-                if (items.length === 0) return null;
-                return (
-                  <div key={group}>
-                    <div className={styles.grp}>{group}</div>
-                    {items.map((item) => (
-                      <button
-                        key={item.id}
-                        type="button"
-                        className={`${styles.ep} ${item.id === selectedId ? styles.epActive : ""}`}
-                        onClick={() => setSelectedId(item.id)}
-                      >
-                        <span className={`${styles.epVerb} ${mode === "mcp" ? styles.epVerbTool : ""}`}>
-                          {mode === "mcp" ? "TOOL" : "POST"}
-                        </span>
-                        <span className={styles.epName}>{item.id}</span>
-                      </button>
-                    ))}
-                  </div>
-                );
-              })}
-            </div>
-          </aside>
-
-          {/* 请求 */}
-          <section className={styles.req}>
-            <div className={styles.reqHead}>
-              <div className={styles.reqRow1}>
-                <span className={styles.verb}>{verb}</span>
-                <span className={styles.path}>{displayPath}</span>
-              </div>
-              <h2 className={styles.reqTitle}>{op.id}</h2>
-              <p className={styles.reqSum}>{op.summary}</p>
-            </div>
-            <div className={styles.reqBody}>
-              {visibleQuery.length > 0 ? (
-                <div className={styles.sec}>
-                  <div className={styles.secHead}>
-                    {mode === "mcp" ? "参数" : "QUERY 参数"} <span className={styles.cnt}>{visibleQuery.length}</span>
-                  </div>
-                  <div className={styles.qp}>
-                    {visibleQuery.map((param) => (
-                      <QueryParamRow
-                        key={param.name}
-                        param={param}
-                        locked={param.name === "kn_id"}
-                        value={param.name === "kn_id" ? knId : queryVals[param.name] ?? param.value}
-                        onChange={(value) => setQueryVals((prev) => ({ ...prev, [param.name]: value }))}
-                      />
-                    ))}
-                  </div>
-                </div>
-              ) : null}
-
-              {op.body !== null ? (
-                <div className={styles.sec}>
-                  <div className={styles.secHead}>
-                    请求体 <span className={styles.sub}>application/json</span>
-                  </div>
-                  <div className={styles.editor}>
-                    <div className={styles.editbar}>
-                      <span className={styles.editLbl}>body.json</span>
-                      <button
-                        type="button"
-                        className={styles.mini}
-                        onClick={() => {
-                          try {
-                            setBodyText(JSON.stringify(JSON.parse(bodyText), null, 2));
-                            setBodyError(null);
-                          } catch (error) {
-                            setBodyError(error instanceof Error ? error.message : "JSON 解析失败");
-                          }
-                        }}
-                      >
-                        格式化
-                      </button>
-                    </div>
-                    <JsonEditor value={bodyText} onChange={setBodyText} />
-                    {bodyError ? <div className={styles.bodyErr}>{bodyError}</div> : null}
-                  </div>
-                </div>
-              ) : null}
-
-              {mode === "mcp" ? (
-                <div className={styles.sec}>
-                  <div className={styles.secHead}>
-                    Schema <span className={styles.sub}>tools/list</span>
-                    {toolDefs || toolsError ? (
-                      <button type="button" className={styles.mini} onClick={() => loadTools(true)}>
-                        刷新
-                      </button>
-                    ) : null}
-                  </div>
-                  {toolsLoading ? (
-                    <div className={styles.schemaHint}>
-                      <Spin size="small" /> 拉取工具 schema…
-                    </div>
-                  ) : toolsError ? (
-                    <div className={styles.schemaHint}>加载失败：{toolsError}</div>
-                  ) : currentTool ? (
-                    <>
-                      <CodeBlock
-                        title="Input Schema"
-                        code={JSON.stringify(currentTool.inputSchema ?? {}, null, 2)}
-                        json
-                        onCopy={() => copy(JSON.stringify(currentTool.inputSchema ?? {}, null, 2), "Input Schema 已复制")}
-                      />
-                      {currentTool.outputSchema !== undefined ? (
-                        <CodeBlock
-                          title="Output Schema"
-                          code={JSON.stringify(currentTool.outputSchema, null, 2)}
-                          json
-                          onCopy={() => copy(JSON.stringify(currentTool.outputSchema, null, 2), "Output Schema 已复制")}
-                        />
-                      ) : (
-                        <div className={styles.schemaHint}>后端未在 tools/list 提供 Output Schema。</div>
-                      )}
-                    </>
-                  ) : toolDefs ? (
-                    <div className={styles.schemaHint}>tools/list 未包含「{op.id}」。</div>
-                  ) : null}
-                </div>
-              ) : null}
-            </div>
-            <div className={styles.actions}>
-              <button type="button" className={styles.sendReq} onClick={() => void onSend()} disabled={sending}>
-                {sending ? <Spin size="small" /> : null}
-                发送请求
-              </button>
-              <button type="button" className={styles.resetBtn} onClick={() => setBodyText(exampleBodyText(op, mode, knId))}>
-                恢复示例
-              </button>
-              {opSupportsTestData(op.id) ? (
-                <Tooltip title="用当前网络真实 schema + 样本行填充，可直接发送">
-                  <button
-                    type="button"
-                    className={styles.testBtn}
-                    onClick={() => void onFillTestData()}
-                    disabled={fillingTest}
-                  >
-                    {fillingTest ? <Spin size="small" /> : <ThunderboltFilled />} 填充测试数据
-                  </button>
-                </Tooltip>
-              ) : null}
-              <button type="button" className={styles.dataBtn} onClick={() => setRightTab("data")}>
-                <DatabaseOutlined /> 数据浏览器
-              </button>
-              <span className={styles.kbd}>⌘ + ↵ 发送</span>
-            </div>
-          </section>
-
-          {/* 响应 / 数据浏览器（标签切换；两个视图常驻不卸载，切换不丢预览/筛选上下文） */}
-          <section className={styles.res}>
-            <div className={styles.rightTabs}>
-              <button
-                type="button"
-                className={`${styles.rightTab} ${rightTab === "res" ? styles.rightTabOn : ""}`}
-                onClick={() => setRightTab("res")}
-              >
-                响应
-              </button>
-              <button
-                type="button"
-                className={`${styles.rightTab} ${rightTab === "data" ? styles.rightTabOn : ""}`}
-                onClick={() => setRightTab("data")}
-              >
-                <DatabaseOutlined /> 数据浏览器
-              </button>
-            </div>
-            <div className={`${styles.rightView} ${rightTab === "res" ? "" : styles.rightHidden}`}>
-            <div className={styles.resHead}>
-              <span className={styles.resTitle}>响应</span>
-              {response ? (
-                <>
-                  <span className={`${styles.pill} ${response.ok ? styles.pillOk : styles.pillErr}`}>
-                    <span className={styles.pillDot} />
-                    {response.status} {response.statusText}
-                  </span>
-                  <span className={styles.resMeta}>
-                    {response.latencyMs}ms · {formatBytes(response.sizeBytes)}
-                  </span>
-                  <button
-                    type="button"
-                    className={styles.copyResp}
-                    onClick={() => copy(responseView?.text ?? response.text, "响应已复制")}
-                  >
-                    <CopyOutlined /> 复制结果
-                  </button>
-                </>
-              ) : (
-                <span className={styles.resHint}>尚未发送请求</span>
-              )}
-            </div>
-            <div className={styles.resBody}>
-              {sending ? (
-                <div className={styles.resEmpty}>
-                  <Spin />
-                </div>
-              ) : reqError ? (
-                <div className={styles.resError}>
-                  <ApiOutlined />
-                  <div>
-                    <strong>请求失败</strong>
-                    <p>{reqError}</p>
-                  </div>
-                </div>
-              ) : responseView ? (
-                <pre className={styles.out}>
-                  {responseView.kind === "toon" ? (
-                    <>
-                      <span className={styles.toonTag}>TOON</span>
-                      {responseView.text}
-                    </>
-                  ) : (
-                    <JsonHighlight text={responseView.text} />
-                  )}
-                </pre>
-              ) : (
-                <div className={styles.resEmpty}>
-                  <ApiOutlined className={styles.resEmptyIc} />
-                  <h3>准备就绪</h3>
-                  <p>选择接口、确认 kn_id 与参数后点击「发送请求」查看实时响应。</p>
-                </div>
-              )}
-            </div>
-            <div className={`${styles.curl} ${curlOpen ? styles.curlOpen : ""}`}>
-              <div className={styles.curlHead} onClick={() => setCurlOpen((value) => !value)}>
-                <span className={styles.curlLbl}>
-                  <span className={styles.chev}>▶</span> cURL
-                </span>
-                <button
-                  type="button"
-                  className={styles.mini}
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    copy(curl, "cURL 已复制");
-                  }}
-                >
-                  复制
-                </button>
-              </div>
-              {curlOpen ? (
-                <div className={styles.curlBody}>
-                  <pre className={styles.curlPre}>{curl}</pre>
-                </div>
-              ) : null}
-            </div>
-            </div>
-            <div className={`${styles.rightView} ${rightTab === "data" ? "" : styles.rightHidden}`}>
-              <DataBrowserPanel
-                active={rightTab === "data"}
-                env={env}
-                knName={network?.name ?? ""}
-                onFillField={fillBodyField}
-                onFillResource={fillResource}
-                onFillConceptGroup={fillConceptGroup}
-                onFillTest={opFillsFromObjectType ? fillTestFromObjectType : undefined}
-                onFillRelation={op.id === "query_instance_subgraph" ? fillSubgraphFromRelation : undefined}
-                copy={copy}
-                auth={tokenProvider}
-              />
-            </div>
-          </section>
-        </div>
+        <ContextLoaderIntegrationPanel
+          mode={mode}
+          knId={knId}
+          activeOps={activeOps}
+          op={op}
+          selectedId={selectedId}
+          onSelectOp={setSelectedId}
+          filter={filter}
+          onFilterChange={setFilter}
+          visibleQuery={visibleQuery}
+          queryVals={queryVals}
+          onQueryChange={(name, value) => setQueryVals((prev) => ({ ...prev, [name]: value }))}
+          bodyText={bodyText}
+          onBodyTextChange={setBodyText}
+          bodyError={bodyError}
+          onFormatBody={() => {
+            try {
+              setBodyText(JSON.stringify(JSON.parse(bodyText), null, 2));
+              setBodyError(null);
+            } catch (error) {
+              setBodyError(error instanceof Error ? error.message : "JSON ????");
+            }
+          }}
+          displayPath={displayPath}
+          response={response}
+          responseView={responseView}
+          reqError={reqError}
+          sending={sending}
+          onSend={() => void onSend()}
+          onResetBody={() => setBodyText(exampleBodyText(op, mode, knId))}
+          fillingTest={fillingTest}
+          onFillTestData={() => void onFillTestData()}
+          rightTab={rightTab}
+          onRightTabChange={setRightTab}
+          curlOpen={curlOpen}
+          onCurlOpenChange={setCurlOpen}
+          curl={curl}
+          onCopy={copy}
+          onOpenGuide={() => setGuideOpen(true)}
+          onOpenDiscover={() => setDiscoverOpen(true)}
+          toolDefs={toolDefs}
+          toolsLoading={toolsLoading}
+          toolsError={toolsError}
+          currentTool={currentTool}
+          onReloadTools={() => loadTools(true)}
+          dataBrowserPanel={
+            <DataBrowserPanel
+              active={rightTab === "data"}
+              env={env}
+              knName={network?.name ?? ""}
+              onFillField={fillBodyField}
+              onFillResource={fillResource}
+              onFillConceptGroup={fillConceptGroup}
+              onFillTest={opFillsFromObjectType ? fillTestFromObjectType : undefined}
+              onFillRelation={op.id === "query_instance_subgraph" ? fillSubgraphFromRelation : undefined}
+              copy={copy}
+              auth={tokenProvider}
+            />
+          }
+        />
       )}
 
       <McpSetupModal
@@ -1651,37 +746,5 @@ export function ExperienceScene() {
         copy={copy}
       />
     </section>
-  );
-}
-
-function QueryParamRow({
-  param,
-  value,
-  locked,
-  onChange,
-}: {
-  param: ContextLoaderOp["query"][number];
-  value: string;
-  locked: boolean;
-  onChange: (value: string) => void;
-}) {
-  return (
-    <>
-      <div className={styles.qpKey}>
-        {param.name}
-        {param.required ? <span className={styles.star}>*</span> : null}
-      </div>
-      {param.options ? (
-        <Select
-          className={styles.qpSelect}
-          value={value}
-          onChange={(next) => onChange(next)}
-          options={param.options.map((option) => ({ value: option, label: option }))}
-          popupMatchSelectWidth={false}
-        />
-      ) : (
-        <input className={styles.qpInput} value={value} disabled={locked} onChange={(e) => onChange(e.target.value)} />
-      )}
-    </>
   );
 }
